// app/interfaces/MainComponent/OrganizeSection.ts
export interface IOrganizeSection {
    _id: string;
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}

export interface IOrganizeSectionInput {
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}